fix: open deployment selector in conversation
