+++ apps/chat/src/components/ConversationView/ConversationView.tsx
@@ -38,11 +38,9 @@ import {
   BasicI18nKeys,
   ButtonsI18nKeys,
   ChatI18nKeys,
-  DeploymentSelectorI18nKeys,
   ConversationI18nKeys,
   ConversationPanelI18nKeys,
   DialFileManagerI18nKeys,
-  FavoritesI18nKeys,
   FileDndI18nKeys,
 } from '../../constants/translation-keys';
 import { useUser } from '../../context/auth/UserContext';
@@ -55,18 +53,19 @@ import { useChatSettingsFormConfig } from '../../hooks/conversation/useChatSetti
 import { useConversationScroll } from '../../hooks/conversation/useConversationScroll';
 import { useModelSelectorLabels } from '../../hooks/conversation/useModelSelectorLabels';
 import { useKeyboardShortcutPreference } from '../../hooks/keyboard-shortcut/useKeyboardShortcutPreference';
-import useFavoriteApplications from '../../hooks/useFavoriteApplications/useFavoriteApplications';
 import { usePageFileDrag } from '../../hooks/usePageFileDrag';
 import {
   dialFilesToAttachments,
   dialFolderPathToAttachment,
 } from '../../utils/dial-file-to-attachment';
 import { resolveCatalogIconUrl } from '../../utils/icon-path';
-import { mapDeploymentToCatalogItem } from '../../utils/map-deployment-to-catalog-item';
 import { isMessageChanged } from '../../utils/message-utils';
-import DeploymentSelectorPanel from '../DeploymentSelector/DeploymentSelectorPanel';
 import type { AttachResult } from '../DialFileManagerModal/types/attach-result';
 import ConversationMessageItem from './ConversationMessageItem';
+
+const DeploymentSelectorOverlay = lazy(
+  () => import('../DeploymentSelector/DeploymentSelectorOverlay'),
+);
 
 const ConversationInput = lazy(async () => {
   const module = await import('@epam/ai-dial-conversation-input');
@@ -116,8 +115,6 @@ interface Props {
   onTranscribeAudio?: (audioUrl: string) => Promise<string>;
   conversation: Conversation;
   onConversationChange: (conv: Conversation) => void;
-  /** Called when the user clicks "Browse full catalog" inside the model picker. */
-  onBrowseCatalog?: () => void;
   /**
    * When provided, the model selector shows this model only and renders
    * disabled (dimmed, does not open) instead of allowing a different model
@@ -155,7 +152,6 @@ const ConversationView: FC<Props> = ({
   onTranscribeAudio,
   conversation,
   onConversationChange,
-  onBrowseCatalog,
   fixedModel,
 }) => {
   const isModelFixed = !!fixedModel;
@@ -181,28 +177,11 @@ const ConversationView: FC<Props> = ({
     isLoading,
     error,
   } = useDeployments();
-  const { favoriteIds, toggleFavorite } = useFavoriteApplications();
   const activeDeploymentId = fixedModel?.id ?? selectedItemId;
-
-  const favoriteCatalogItems = useMemo(
-    () =>
-      items
-        .filter((d) => favoriteIds.has(d.id))
-        .map((d) => mapDeploymentToCatalogItem(d, favoriteIds)),
-    [items, favoriteIds],
-  );
 
   const selectedDeployment = useMemo(
     () => items.find((item) => item.id === activeDeploymentId),
     [items, activeDeploymentId],
-  );
-
-  const selectedCatalogItem = useMemo(
-    () =>
-      selectedDeployment
-        ? mapDeploymentToCatalogItem(selectedDeployment, favoriteIds)
-        : undefined,
-    [selectedDeployment, favoriteIds],
   );
 
   const {
@@ -681,32 +660,9 @@ const ConversationView: FC<Props> = ({
                   isModelFixed
                     ? undefined
                     : (onClose) => (
-                        <DeploymentSelectorPanel
-                          favorites={favoriteCatalogItems}
-                          selectedId={selectedItemId}
-                          selectedItem={selectedCatalogItem}
-                          onSelect={setSelectedItemId}
-                          onToggleFavorite={toggleFavorite}
-                          onBrowseCatalog={onBrowseCatalog}
-                          onClose={onClose}
-                          labels={{
-                            searchPlaceholder: t(
-                              DeploymentSelectorI18nKeys.SearchPlaceholder,
-                            ),
-                            favoritesLabel: t(FavoritesI18nKeys.FavoritesLabel),
-                            emptyHint: t(DeploymentSelectorI18nKeys.EmptyHint),
-                            browseCatalogLabel: t(ButtonsI18nKeys.Browse),
-                            removeFromFavoritesLabel: t(
-                              FavoritesI18nKeys.RemoveFromFavorites,
-                            ),
-                            currentlySelectedLabel: t(
-                              DeploymentSelectorI18nKeys.CurrentlySelectedLabel,
-                            ),
-                            addToFavoritesLabel: t(
-                              FavoritesI18nKeys.AddToFavorites,
-                            ),
-                          }}
-                        />
+                        <Suspense fallback={null}>
+                          <DeploymentSelectorOverlay onClose={onClose} />
+                        </Suspense>
                       )
                 }
               />
